* commit '2f8818c2369f15247b6c610e64926f9d6b1c3c39':
  [BugFix]WeightCompositの戻り値をunsigendからdoubleに変更
  [Rename] IDDFSって言うよりDLS(Depth-Limited) Seachじゃね?
  [Optimize] optimized Structure

+++ Hal/Heuristics/WeightComposit.h
@@ -3,16 +3,16 @@
 #include "../Structures.h"
 #include <map>
 
-class WeightComposit : public Heuristics<unsigned,Field>{
+class WeightComposit : public Heuristics<double,Field>{
 private:
 	typedef Heuristics<unsigned,Field> h_type;
-	std::map<h_type*,float> h_list;
+	std::map<h_type*,double> h_list;
 
 public:
 
-	void AddHeuristic(h_type* heuristics,float weight);
+	void AddHeuristic(h_type* heuristics,double weight);
 
-	unsigned Execution(const Field& field);
+	double Execution(const Field& field);
 
 	WeightComposit();
 };
